// online_learning_system_frontend_dashboard/src/app/shared/models/common/socket-event-type.enum.ts
export enum SocketEventType {
  CONNECT = 'CONNECT',
  START_DRAWING = 'START_DRAWING',
  START_MCQ = 'START_MCQ',
  START_YOUTUBE = 'START_YOUTUBE',
  START_SCORM_OBJECT = 'START_SCORM_OBJECT',
  START_PDF = 'START_PDF',
  START_PRESENTATION = 'START_PRESENTATION',
  START_WEB_PAGE = 'START_WEB_PAGE',
  START_IMAGE = 'START_IMAGE',
  DRAWING_RESULTS = 'DRAWING_RESULTS',
  OTHER_CONTENT_RESULT = 'OTHER_CONTENT_RESULT',
  MCQ_RESULT = 'MCQ_RESULT',
  PAIRED_SUCCESS = 'PAIRED_SUCCESS',
  CLOSE = 'CLOSE'
}